requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.2
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cmaes==0.10.0
cmake==3.27.1
colorama==0.4.6
colorlog==6.7.0
contourpy==1.1.0
croniter==1.4.1
cycler==0.11.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.1
docopt==0.6.2
einops==0.6.1
fastapi==0.101.0
ffmpeg-python==0.2.0
filelock==3.12.2
Flask==2.3.2
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
greenlet==2.0.2
h11==0.14.0
huggingface-hub==0.16.4
HyperPyYAML==1.2.1
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
julius==0.2.7
kiwisolver==1.4.4
lazy_loader==0.3
librosa==0.10.0.post2
lightning==2.0.6
lightning-cloud==0.5.37
lightning-utilities==0.9.0
lit==16.0.6
llvmlite==0.40.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@b91c907694f96a3fb9da03d4bbdc83fbcd3a40a4
optuna==3.3.0
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pipreqs==0.4.13
pooch==1.6.0
primePy==1.3
protobuf==4.23.4
psutil==5.9.5
pyannote.audio @ git+https://github.com/pyannote/pyannote-audio@37b39b04bd4ec63d540a3a8084f0910c4138faf0
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.6
pytorch-metric-learning==2.3.0
pytz==2023.3
PyYAML==6.0.1
readchar==4.0.5
regex==2023.6.3
requests==2.31.0
rich==13.5.2
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
semver==3.0.1
sentencepiece==0.1.99
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
speechbrain==0.5.15
SQLAlchemy==2.0.19
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2
threadpoolctl==3.2.0
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.0.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.2
torchmetrics==1.0.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
triton==2.0.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
whisper==1.1.10
yarg==0.1.9
yarl==1.9.2
